Move env logger to separate subpath export
